Audio preloads metadata now and ends correctly
The 'preload' attribute of the <audio> tags used for previews of tracks
has been changed from 'none' to 'metadata'. This is the recommended
setting on MDN and results in a quicker load time of audio without
completely loading every audio element.

The Javascript code that handles playing and pausing track previews
has been changed as well. The code in general has been cleaned up, and
an event listener has been added for the <audio>'s 'ended' event.